requirements.txt
dj-database-url==0.4.2
dj-static==0.0.6
Django==2.1.7
python-decouple==3.1
static3==0.7.0
django-extensions==2.0.7
django-test-without-migrations==0.6
djangorestframework==3.9.2
django-cors-headers==2.4.1
gunicorn==19.4.1
psycopg2==2.7.*